hasItPumped/backend/pyproject.toml
[tool.poetry]
name = "solana-token-api"
version = "1.0.0"
description = "Solana Token Analysis API"
authors = ["Matthew Carter <matthewcarter2301@gmail.com>"]
readme = "README.md"
packages = [
  { include = "solana_token_api", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = { extras = ["standard"], version = "^0.22.0" }
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
xgboost = "^1.7.5"
requests = "^2.28.0"
scikit-learn = "^1.6.1"
setuptools = "^80.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.3.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
multi_line_output = 3